fix sighash of coinbase transaction
bitcore-lib-cash was not able to compute the sighash of a coinbase
transaction. This is understandable because wallets built on bitcore
would probably not normally need to compute the sighash of a coinbase
transaction, which is something only miners and full nodes need to do.
Nonetheless, there are reasons why someone might want to use bitcore to
compute the sighash of a coinbase transaction, such as computing the
normalized transaction id of all transactions on the blockchain.

Coinbase transactions have "null" inputs, which means the input has 0
for the txid and max int for the output number. The bitcore input
.script property returns javascript null in that case, however the input
setScript property expects a script. A simple way to resolve this is to
allow the setScript method to accept a null value and to interpret that
as an empty script.